+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- Raise an error if too few conformers survive filtering in [38](https://github.com/cole-group/presto/pull/38). Fixes [#30](https://github.com/cole-group/presto/issues/30).
 - Remove stereochemical information from generated types in [#36](https://github.com/cole-group/presto/pull/37). This avoids a niche issue where mixing the RDKit and OpenEye toolkits would result in failed type generation for e.g. chiral sulfoxides.
 
 ### Maintenance